made voting and marking a task as done thread safe

+++ src/main/java/com/example/vote/VoteService.java
@@ -8,12 +8,17 @@ import com.example.upload.entity.UploadId;
 import com.example.upload.repo.UploadRepository;
 import com.example.user.entity.User;
 import com.example.user.repository.UserRepository;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.OptimisticLockException;
+import jakarta.persistence.PersistenceContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.concurrent.ConcurrentHashMap;
 
 @Service
 @Transactional
@@ -22,6 +27,9 @@ public class VoteService {
     private final UploadRepository uploadRepository;
     private final TaskService taskService;
     private final UserRepository userRepository;
+    private final ConcurrentHashMap<UploadId, Object> uploadLocks = new ConcurrentHashMap<>();
+    @PersistenceContext
+    private EntityManager entityManager;
     @Autowired
     public VoteService(TaskRepository taskRepository,UploadRepository uploadRepository,TaskService taskService,UserRepository userRepository){
         this.taskRepository=taskRepository;
@@ -34,23 +42,33 @@ public class VoteService {
         return userRepository.findByUsername(authentication.getName())
                 .orElseThrow(() -> new AccessDeniedException("User not found"));
     }
-    public void voteFor(Long votedUserId,Long task_id){
-        Task task=taskRepository.findById(task_id).orElseThrow(()->new IllegalArgumentException("task not found"));
-        User user=getCurrentUser();
+    //we shouldn't allow user to send vote requests for himself to implement later!
+    //for simplicity now , we allow the same user to vote for same file more than once
+    public void voteFor(Long votedUserId, Long taskId) {
+        Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new IllegalArgumentException("task not found"));
 
-        if(!task.getGoal().getMembers().contains(user)){
+        User user = getCurrentUser();
+
+        if (!task.getGoal().getMembers().contains(user)) {
             throw new AccessDeniedException("not authorized to vote for this file");
         }
-        if(uploadRepository.findById(new UploadId(votedUserId,task_id)).isEmpty()){
+
+        UploadId key = new UploadId(votedUserId, taskId);
+
+        if (uploadRepository.findById(key).isEmpty()) {
             throw new IllegalArgumentException("no uploads exist for this user");
         }
 
-        uploadRepository.incrementVotes(votedUserId,task_id);
-        Upload upload=uploadRepository.findById(new UploadId(votedUserId,task_id))
-                .orElseThrow(()->new IllegalArgumentException("no uploads exist for this user and task"));
-        if(upload.getCurrentVotes()>=(task.getGoal().getVotesToMarkCompleted())){
+        int isUploadedSuccessfully=uploadRepository.incrementVotesIfBelowThreshold(votedUserId, taskId,task.getGoal().getVotesToMarkCompleted());
+
+        entityManager.clear();
+
+        Upload upload=uploadRepository.findById(key).orElseThrow();
+        if(isUploadedSuccessfully==1 && upload.getCurrentVotes()==task.getGoal().getVotesToMarkCompleted()){
             uploadRepository.delete(upload);
-            taskService.markDone(task_id);
+            taskService.markDone(taskId,votedUserId);
         }
+
     }
 }
